+++ src/examples/02.custom-formatter.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2020 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,22 +14,22 @@
  *  limitations under the License.
  */
 
-import Loggers from "../lib";
+import Loggers from '../lib';
 
 (function customFormatter(): void {
 
     /**
      * First, create a log controller, giving a subject and a formater.
      */
-    let logs1 = Loggers.createTextLogger(
-        "Custom-Formatter 1",
+    const logs1 = Loggers.createTextLogger(
+        'Custom-Formatter 1',
         function(log, subj, lv, dt, traces): string {
 
             if (traces) {
 
                 return `${dt.toISOString()} - ${subj} - ${lv} - ${log}
 
-  ${traces.join("\n  ")}
+  ${traces.join('\n  ')}
 `;
             }
 
@@ -47,22 +47,22 @@ import Loggers from "../lib";
     /**
      * Output a log of INFO level.
      */
-    logs1.info("This is INFO log.");
+    logs1.info('This is INFO log.');
 
-    logs1.error("This is ERROR log.");
+    logs1.error('This is ERROR log.');
 
     /**
      * Now the DEBUG logs couldn't be output (No errors, but ignored.)
      */
-    logs1.debug("This is DEBUG log.");
+    logs1.debug('This is DEBUG log.');
 
     logs1.enableTrace();
 
-    logs1.warning("This is WARNING log.");
+    logs1.warning('This is WARNING log.');
 
     logs1.enableTrace(10);
 
-    logs1.notice("This is NOTICE log.");
+    logs1.notice('This is NOTICE log.');
 
     /**
      * Or used a pre-registered formatter.
@@ -72,14 +72,14 @@ import Loggers from "../lib";
      * First, create a log controller, giving a subject and a formater.
      */
     Loggers.registerTextFormatter(
-        "custom_text_formatter",
+        'custom_text_formatter',
         function(log, subj, lv, dt, traces): string {
 
             if (traces) {
 
                 return `${dt.toISOString()} - ${subj} - ${lv} - ${log}
 
-  ${traces.join("\n  ")}
+  ${traces.join('\n  ')}
 `;
             }
 
@@ -87,9 +87,9 @@ import Loggers from "../lib";
         }
     );
 
-    let logs2 = Loggers.createTextLogger(
-        "Custom-Formatter 2",
-        "custom_text_formatter"
+    const logs2 = Loggers.createTextLogger(
+        'Custom-Formatter 2',
+        'custom_text_formatter'
     );
 
     /**
@@ -102,20 +102,20 @@ import Loggers from "../lib";
     /**
      * Output a log of INFO level.
      */
-    logs2.info("This is INFO log.");
+    logs2.info('This is INFO log.');
 
-    logs2.error("This is ERROR log.");
+    logs2.error('This is ERROR log.');
 
     /**
      * Now the DEBUG logs couldn't be output (No errors, but ignored.)
      */
-    logs2.debug("This is DEBUG log.");
+    logs2.debug('This is DEBUG log.');
 
     logs2.enableTrace();
 
-    logs2.warning("This is WARNING log.");
+    logs2.warning('This is WARNING log.');
 
     logs2.enableTrace(10);
 
-    logs2.notice("This is NOTICE log.");
+    logs2.notice('This is NOTICE log.');
 })();