+++ src/lib/Drivers/ColorfulTTY.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2020 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -16,15 +16,15 @@
 
 // tslint:disable:no-console
 
-import { IDriver } from "../Common";
+import { IDriver } from '../Common';
 
-export type ForeColorSet = "blue" | "cyan" | "green" | "magenta" | "grey" |
-                            "red" | "yellow" | "white" | "black" | "default";
+export type ForeColorSet = 'blue' | 'cyan' | 'green' | 'magenta' | 'grey' |
+'red' | 'yellow' | 'white' | 'black' | 'default';
 
 export type BgColorSet = ForeColorSet;
 
-const BG_COLOR_ENDING = "\x1B[49m";
-const FORE_COLOR_ENDING = "\x1B[39m";
+const BG_COLOR_ENDING = '\x1B[49m';
+const FORE_COLOR_ENDING = '\x1B[39m';
 
 type Writer = (
     text: string,
@@ -33,44 +33,39 @@ type Writer = (
     date: Date
 ) => void;
 
-function NON_COLORFUL_WRITER(
-    text: string,
-    subject: string,
-    level: string,
-    date: Date
-): void {
+function NON_COLORFUL_WRITER(text: string): void {
 
     console.log(text);
 }
 
 const FORE_COLORS: Record<ForeColorSet, string> = {
-    "default": "",
-    "blue": "\u001b[34m",
-    "cyan": "\u001b[36m",
-    "green": "\u001b[32m",
-    "magenta": "\u001b[35m",
-    "red": "\u001b[31m",
-    "yellow": "\u001b[33m",
-    "white": "\u001b[37m",
-    "grey": "\u001b[90m",
-    "black": "\u001b[30m"
+    'default': '',
+    'blue': '\u001b[34m',
+    'cyan': '\u001b[36m',
+    'green': '\u001b[32m',
+    'magenta': '\u001b[35m',
+    'red': '\u001b[31m',
+    'yellow': '\u001b[33m',
+    'white': '\u001b[37m',
+    'grey': '\u001b[90m',
+    'black': '\u001b[30m'
 };
 
 const BG_COLORS: Record<BgColorSet, string> = {
-    "default": "",
-    "white": "\u001b[47m",
-    "grey": "\u001b[49;5;8m",
-    "black": "\u001b[40m",
-    "blue": "\u001b[44m",
-    "cyan": "\u001b[46m",
-    "green": "\u001b[42m",
-    "magenta": "\u001b[45m",
-    "red": "\u001b[41m",
-    "yellow": "\u001b[43m"
+    'default': '',
+    'white': '\u001b[47m',
+    'grey': '\u001b[49;5;8m',
+    'black': '\u001b[40m',
+    'blue': '\u001b[44m',
+    'cyan': '\u001b[46m',
+    'green': '\u001b[42m',
+    'magenta': '\u001b[45m',
+    'red': '\u001b[41m',
+    'yellow': '\u001b[43m'
 };
 
 export interface IColorfulTTYDriver
-extends IDriver {
+    extends IDriver {
 
     /**
      * Set the fore-color for the logs of a level.
@@ -103,7 +98,7 @@ interface IStyle {
  *
  * @see https://github.com/Microsoft/TypeScript/issues/24587
  */
-const DEFAULT_LEVEL: string = Symbol("__default__") as any;
+const DEFAULT_LEVEL: string = Symbol('__default__') as any;
 
 class ColorfulTTYDriver
 implements IColorfulTTYDriver {
@@ -117,26 +112,26 @@ implements IColorfulTTYDriver {
     public constructor() {
 
         this._bgColors = {
-            [DEFAULT_LEVEL]: "default"
+            [DEFAULT_LEVEL]: 'default'
         };
         this._foreColors = {
-            [DEFAULT_LEVEL]: "default"
+            [DEFAULT_LEVEL]: 'default'
         };
 
         this._levels = {
             [DEFAULT_LEVEL]: {
-                start: "",
-                end: ""
+                start: '',
+                end: ''
             }
         };
 
-        this.write = this._buildWriter();
+        this.write = this._buildWriter() as any;
     }
 
     public bgColor(color: BgColorSet, level?: string): this {
 
-        this._bgColors[level || DEFAULT_LEVEL] = color === "default" ?
-            "" : BG_COLORS[color];
+        this._bgColors[level || DEFAULT_LEVEL] = color === 'default' ?
+            '' : BG_COLORS[color];
 
         this._rebuild(level || DEFAULT_LEVEL);
 
@@ -145,8 +140,8 @@ implements IColorfulTTYDriver {
 
     public foreColor(color: ForeColorSet, level?: string): this {
 
-        this._foreColors[level || DEFAULT_LEVEL] = color === "default" ?
-            "" : FORE_COLORS[color];
+        this._foreColors[level || DEFAULT_LEVEL] = color === 'default' ?
+            '' : FORE_COLORS[color];
 
         this._rebuild(level || DEFAULT_LEVEL);
 
@@ -155,30 +150,25 @@ implements IColorfulTTYDriver {
 
     private _rebuild(level: string/* | symbol*/): void {
 
-        let start: string = "";
-        let end: string = "";
+        let start: string = '';
+        let end: string = '';
 
         if (this._foreColors[level]) {
 
-            start += this._foreColors[level] || "";
+            start += this._foreColors[level] || '';
             end = FORE_COLOR_ENDING + end;
         }
 
         if (this._bgColors[level]) {
 
-            start += this._bgColors[level] || "";
+            start += this._bgColors[level] || '';
             end = BG_COLOR_ENDING + end;
         }
 
         this._levels[level] = { start, end };
     }
 
-    public write(
-        text: string,
-        subject: string,
-        level: string,
-        date: Date
-    ): void {
+    public write(): void {
 
         return;
     }
@@ -200,14 +190,14 @@ implements IColorfulTTYDriver {
             // @ts-ignore
             process.stdout.isTTY ||
             // @ts-ignore
-            process.stdout.constructor.name === "Socket" // Debugging
+            process.stdout.constructor.name === 'Socket' // Debugging
         );
     }
 
     public static isNodeJS(): boolean {
 
         // @ts-ignore
-        return typeof process === "object" && typeof process.stdout === "object";
+        return typeof process === 'object' && typeof process.stdout === 'object';
     }
 
     private _buildWriter(): Writer {
@@ -224,18 +214,18 @@ implements IColorfulTTYDriver {
 
         const cs: string[] = [];
 
-        cs.push(`return function(text, subject, level, date) {`);
-        cs.push(`const dec = this._levels[level] || this._levels[DEFAULT_LEVEL];`);
-        cs.push(`return console.log(`);
-        cs.push(`    text.split("\\n").map(`);
-        cs.push(`       (x) => \`\${dec.start}\${x}\${dec.end}\``);
-        cs.push(`    ).join("\\n")`);
-        cs.push(`);`);
-        cs.push(`};`);
+        cs.push('return function(text, subject, level, date) {');
+        cs.push('const dec = this._levels[level] || this._levels[DEFAULT_LEVEL];');
+        cs.push('return console.log(');
+        cs.push('    text.split("\\n").map(');
+        cs.push('       (x) => `${dec.start}${x}${dec.end}`');
+        cs.push('    ).join("\\n")');
+        cs.push(');');
+        cs.push('};');
 
         return (new Function(
-            "DEFAULT_LEVEL",
-            cs.join("\n")
+            'DEFAULT_LEVEL',
+            cs.join('\n')
         ))(DEFAULT_LEVEL);
     }
 }
@@ -248,7 +238,7 @@ export function createColorfulTTYDriver(): IColorfulTTYDriver {
     if (!ColorfulTTYDriver.isTerminal()) {
 
         console.warn(
-            "The ColorfulTTYDriver is only usable in Node.JS terminal."
+            'The ColorfulTTYDriver is only usable in Node.JS terminal.'
         );
     }