+++ src/lib/Drivers/Console.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2020 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -16,17 +16,12 @@
 
 // tslint:disable:no-console
 
-import { IDriver } from "../Common";
+import { IDriver } from '../Common';
 
 class ConsoleDriver
 implements IDriver {
 
-    public write(
-        text: string,
-        subject: string,
-        level: string,
-        date: Date
-    ): void {
+    public write(text: string): void {
 
         return console.log(text);
     }