+++ src/lib/Internal.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2020 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -22,10 +22,10 @@ export interface ILevelOptions {
     /**
      * If the level is enabled for output.
      */
-    "enabled": boolean;
+    'enabled': boolean;
 
     /**
      * How many lines of stack trace could be logged.
      */
-    "trace": number;
+    'trace': number;
 }